requirements.txt
aiodns==3.0.0
aiofiles==0.8.0
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
aiosqlite==0.17.0
APScheduler==3.9.1
async-timeout==4.0.2
attrs==21.4.0
Brotli==1.0.9
cchardet==2.1.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
ciso8601==2.2.0
colorlog==6.6.0
frozenlist==1.3.0
hikari==2.0.0.dev109
hikari-lightbulb==2.2.4
hikari-sake==1.0.2a1
idna==3.3
multidict==6.0.2
pycares==4.2.1
pycparser==2.21
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
uvloop==0.16.0
yarl==1.7.2